fixed issue #130
Former-commit-id: 9d2c1938b3a3e1da171ee82c58128a59d10f22f4

+++ Chipmunk/src/chipmunk.h
@@ -26,7 +26,7 @@
 extern "C" {
 #endif
 	
-typedef double cpFloat;
+typedef float cpFloat;
 	
 static inline cpFloat
 cpfmax(cpFloat a, cpFloat b)
